cpu/esp32/esp_wifi: move gnrc specific code to separate file
